requirement.txt
django-REST-framework
beautifulsoup4
requests